Change `UINT32` to `size_t`, and then cast the result.
This will simplify Read/WriteVirtual/PhysicalMemory, and is just the right interface, despite all the callers using a small number.

Related work items: #57987809

+++ MsvmPkg/Library/BdLib/dbgio.c
@@ -52,7 +52,7 @@ Return Value:
     // Move the output string to the message buffer.
     //
 
-    length = BdMoveMemory((PCHAR)BdMessageBuffer,
+    length = (UINT32)BdMoveMemory((PCHAR)BdMessageBuffer,
                              (PCHAR)Output->Buffer,
                              Output->Length);
 
@@ -128,7 +128,7 @@ Return Value:
     // Move the output string to the message buffer.
     //
 
-    length = BdMoveMemory((PCHAR)BdMessageBuffer,
+    length = (UINT32)BdMoveMemory((PCHAR)BdMessageBuffer,
                              (PCHAR)Output->Buffer,
                              Output->Length);
